+++ Source/ProjectFirma.Web/Security/FundingSourceViewFeature.cs
@@ -20,7 +20,6 @@ Source code is available upon request via <support@sitkatech.com>.
 -----------------------------------------------------------------------*/
 
 using ProjectFirma.Web.Models;
-using ProjectFirma.Web.Security;
 using ProjectFirma.Web.Security.Shared;
 
 namespace ProjectFirma.Web.Security
