+++ Source/ProjectFirma.Web/Security/ProjectCreateNewFeature.cs
@@ -27,7 +27,7 @@ namespace ProjectFirma.Web.Security
     public class ProjectCreateNewFeature : FirmaFeature
     {
         public ProjectCreateNewFeature()
-            : base(new List<Role> { Role.SitkaAdmin, Role.Admin })
+            : base(new List<Role> {Role.SitkaAdmin, Role.Admin, Role.ProjectApprover})
         {
         }
     }
