+++ Source/ProjectFirma.Web/Security/ProjectNoteUpdateEditFeature.cs
@@ -29,7 +29,7 @@ namespace ProjectFirma.Web.Security
         private readonly FirmaFeatureWithContextImpl<ProjectNoteUpdate> _firmaFeatureWithContextImpl;
 
         public ProjectNoteUpdateEditFeature()
-            : base(new List<Role> { Role.Normal, Role.SitkaAdmin, Role.Admin })
+            : base(new List<Role> { Role.Normal, Role.SitkaAdmin, Role.Admin, Role.ProjectApprover })
         {
             _firmaFeatureWithContextImpl = new FirmaFeatureWithContextImpl<ProjectNoteUpdate>(this);
             ActionFilter = _firmaFeatureWithContextImpl;
