+++ Source/ProjectFirma.Web/Security/ProposedProjectCreateNewFeature.cs
@@ -27,7 +27,7 @@ namespace ProjectFirma.Web.Security
     public class ProposedProjectCreateNewFeature : FirmaFeature
     {
         public ProposedProjectCreateNewFeature()
-            : base(new List<Role> { Role.Normal, Role.SitkaAdmin, Role.Admin })
+            : base(new List<Role> { Role.Normal, Role.SitkaAdmin, Role.Admin, Role.ProjectApprover })
         {
         }
     }
