+++ Source/ProjectFirma.Web/Views/ProjectUpdate/ProjectUpdateBatchDiffLog.cshtml
@@ -103,7 +103,7 @@ Source code is available upon request via <support@sitkatech.com>.
     <div class="panel panelFirma">
         <div class="panel-heading panelTitle">
             <h2>Notes</h2>
-        </div>>
+        </div>
         @if (ViewDataTyped.ProjectUpdateBatch.NotesDiffLogHtmlString != null)
         {
             @Html.Raw(ViewDataTyped.ProjectUpdateBatch.NotesDiffLogHtmlString)
